Enhanced test environment Added smoke tests Added compose.test.yaml for integration testing Fixed bug with Converter<T,S> All set

+++ struct-api/src/main/java/ru/lexender/icarusdb/struct/core/StructController.java
@@ -6,11 +6,20 @@ import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.log4j.Log4j2;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import ru.lexender.icarusdb.struct.core.dto.StructRequest;
@@ -46,13 +55,17 @@ public class StructController {
      */
     @GetMapping
     @Operation(summary = "Get all structs", description = "Fetch all structs from the database")
-    public Flux<StructResponse> findAllStructs() {
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Found all structs"),
+    })
+    public ResponseEntity<Flux<StructResponse>> findAllStructs() {
         log.info("Fetching all structs");
-        return structService.findAll()
+        return ResponseEntity.ok(structService.findAll()
                 .map(struct -> {
-                    log.debug("Mapping structs to StructResponses: {}", struct);
+                    log.debug("Converting struct: {}", struct);
                     return objectMapper.convertValue(struct, StructResponse.class);
-                });
+                })
+        );
     }
 
     /**
@@ -62,10 +75,14 @@ public class StructController {
      */
     @DeleteMapping
     @Operation(summary = "Delete all structs", description = "Delete all structs from the database")
-    public Mono<Void> deleteAllStructs() {
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "All structs deleted"),
+    })
+    public Mono<ResponseEntity<Void>> deleteAllStructs() {
         log.info("Deleting all structs");
         return structService.deleteAll()
-                .doOnSuccess(aVoid -> log.info("All structs deleted successfully"))
+                .then(Mono.just(ResponseEntity.noContent().<Void>build()))
+                .doOnSuccess(response -> log.debug("All structs deleted successfully"))
                 .doOnError(e -> log.error("Error deleting all structs", e));
     }
 
@@ -77,13 +94,17 @@ public class StructController {
      */
     @PostMapping
     @Operation(summary = "Create a struct", description = "Create a new struct in the database")
-    public Mono<UUID> createStruct(@RequestBody @Parameter(description = "New struct") StructRequest struct) {
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Struct created"),
+            @ApiResponse(responseCode = "400", description = "Invalid struct supplied"),
+    })
+    public Mono<ResponseEntity<UUID>> createStruct(@Valid @RequestBody @Parameter(description = "New struct") StructRequest struct) {
         log.info("Creating new struct: {}", struct);
-        return Mono.just(objectMapper.convertValue(struct, Struct.class))
-                .flatMap(structService::save)
+        return structService.save(objectMapper.convertValue(struct, Struct.class))
                 .map(Struct::getId)
-                .doOnSuccess(id -> log.info("New struct created with ID: {}", id))
-                .doOnError(e -> log.error("Error creating new struct", e));
+                .map(ResponseEntity.status(HttpStatus.CREATED)::body)
+                .doOnSuccess(response -> log.debug("Struct created with ID: {}", response.getBody()))
+                .doOnError(e -> log.error("Error creating struct", e));
     }
 
     /**
@@ -99,14 +120,21 @@ public class StructController {
             @ApiResponse(responseCode = "400", description = "Invalid id supplied"),
             @ApiResponse(responseCode = "404", description = "Struct not found"),
     })
-    public Mono<StructResponse> findStructById(@PathVariable @Parameter(description = "ID of the struct to be fetched") UUID id) {
+    public Mono<ResponseEntity<StructResponse>> findStructById(@PathVariable @Parameter(description = "ID of the struct to be fetched") UUID id) {
         log.info("Fetching struct with ID: {}", id);
         return structService.findById(id)
                 .map(struct -> {
-                    log.debug("Mapping struct to StructResponse: {}", struct);
-                    return objectMapper.convertValue(struct, StructResponse.class);
+                    StructResponse structResponse = objectMapper.convertValue(struct, StructResponse.class);
+                    return ResponseEntity.ok(structResponse);
                 })
-                .doOnSuccess(structResponse -> log.info("Struct found: {}", structResponse))
+                .defaultIfEmpty(ResponseEntity.notFound().build())
+                .doOnSuccess(response -> {
+                    if (response.getStatusCode() == HttpStatus.NOT_FOUND) {
+                        log.debug("Struct with ID {} not found", id);
+                    } else {
+                        log.debug("Struct found: {}", response.getBody());
+                    }
+                })
                 .doOnError(e -> log.error("Error fetching struct with ID: {}", id, e));
     }
 
@@ -118,10 +146,41 @@ public class StructController {
      */
     @DeleteMapping("/{id}")
     @Operation(summary = "Delete a struct by ID", description = "Delete a specific struct from the database by its ID")
-    public Mono<Void> deleteStructById(@PathVariable @Parameter(description = "ID of the struct to be deleted") UUID id) {
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "Struct deleted"),
+            @ApiResponse(responseCode = "400", description = "Invalid id supplied"),
+            @ApiResponse(responseCode = "404", description = "Struct not found"),
+    })
+    public Mono<ResponseEntity<Void>> deleteStructById(@PathVariable @Parameter(description = "ID of the struct to be deleted") UUID id) {
         log.info("Deleting struct with ID: {}", id);
-        return structService.deleteById(id)
-                .doOnSuccess(aVoid -> log.info("Struct with ID {} deleted successfully", id))
-                .doOnError(e -> log.error("Error deleting struct with ID: {}", id, e));
+        return structService.existsById(id).flatMap(exists -> {
+            if (exists) {
+                return structService.deleteById(id)
+                        .then(Mono.just(ResponseEntity.noContent().<Void>build()))
+                        .doOnSuccess(response -> log.debug("Struct with ID {} deleted successfully", id))
+                        .doOnError(e -> log.error("Error deleting struct with ID: {}", id, e));
+            } else {
+                log.debug("Struct to delete with ID {} not found", id);
+                return Mono.just(ResponseEntity.notFound().build());
+            }
+        });
+    }
+
+    /**
+     * Counts all Structs in the database.
+     *
+     * @return a Mono of the count of all Structs
+     */
+    @GetMapping("/count")
+    @Operation(summary = "Count all structs", description = "Count all structs in the database")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Counted all structs"),
+    })
+    public Mono<ResponseEntity<Long>> countStructs() {
+        log.info("Counting all structs");
+        return structService.count()
+                .map(ResponseEntity.ok()::body)
+                .doOnSuccess(response -> log.debug("Counted all structs: {}", response.getBody()))
+                .doOnError(e -> log.error("Error counting all structs", e));
     }
 }
